requirements.txt
pyrogram==2.0.106
sqlalchemy==2.0.49
psycopg2-binary
python-dotenv
apscheduler
google-genai
fastapi
uvicorn
openpyxl
